requirements.txt
Flask==2.3.3
python-docx==0.8.11
PyPDF2==3.0.1
beautifulsoup4==4.12.2
openai==1.8.4       # Optional for embeddings
